requirements.txt
beautifulsoup4==4.5.1
click==6.6
cookies==2.2.1
Django==1.10
dotenv==0.0.5
Flask==0.11.1
folium==0.2.1
geopy==1.11.0
gevent==1.1.2
google-api-python-client==1.5.3
googlefinance==0.7
greenlet==0.4.10
heroku==0.1.4
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.1
oauth2==1.9.0.post1
oauth2client==3.0.0
ordereddict==1.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==1.5
python-dotenv==0.6.0
python-forecastio==1.3.4
pytz==2016.6.1
requests==2.11.1
requests-file==1.4
responses==0.5.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
slacker==0.9.25
slacksocket==0.7
uritemplate==0.6
websocket==0.2.1
websocket-client==0.37.0
Werkzeug==0.11.11
yelp==1.0.2
